Re-arrange the UI to fit 2 messages
