+++ i18n/qqq.json
@@ -83,8 +83,5 @@
 	"privacy-terms-header": "The header title shown in the Privacy and Terms page",
 	"softkey-consent-agree": "Option to agree to the personal data and usage consent.",
 	"softkey-consent-terms": "Option to go to the application's terms of service.",
-	"softkey-consent-policy": "Option to go to the application's privacy policy.",
-	"consent-title": "Title line of the consent screen.",
-	"consent-subtitle": "Subtitle line of the consent screen.",
-	"consent-message": "Main message of the consent screen."
+	"softkey-consent-policy": "Option to go to the application's privacy policy."
 }
