+++ src/utils/index.js
@@ -4,6 +4,7 @@ export * from './articleHistory'
 export * from './cachedFetch'
 export * from './canonicalizeTitle'
 export * from './confirm'
+export * from './consent'
 export * from './featureFlags'
 export * from './goto'
 export * from './interwikiKeys'
@@ -15,4 +16,3 @@ export * from './onboarding'
 export * from './sendErrorLog'
 export * from './sendEvent'
 export * from './viewport'
-export * from './usageReportConsent'
