+++ style/consent.less
@@ -1,0 +1,22 @@
+.consent {
+	.header {
+		height: @pageHeaderHeight;
+		background-color: #d8d8d8;
+		padding: 3px 0;
+		font-size: 18px;
+		text-align: center;
+	}
+
+	.body {
+		background-color: #fff;
+		height: @availableHeight;
+		text-align: center;
+
+		.message {
+			font-size: 14px;
+			line-height: 1.5;
+			color: #72777d;
+			padding: 10px 10px 0 10px;
+		}
+	}
+}
