+++ style/style.less
@@ -17,6 +17,7 @@
 @import './aboutwikipedia';
 @import './article';
 @import './articlelanguage';
+@import './consent';
 @import './error';
 @import './footer';
 @import './gallery';
@@ -33,4 +34,3 @@
 @import './table';
 @import './textsize';
 @import './toc';
-@import './usageconsent';
