--- style/usageconsent.less
@@ -1,54 +1,0 @@
-.usageconsent {
-	.header {
-		height: @pageHeaderHeight;
-		max-height: @pageHeaderHeight;
-		background-color: #d8d8d8;
-		box-sizing: border-box;
-		padding: 3px 0;
-		font-size: 17px;
-		text-align: center;
-	}
-
-	.body {
-		background-color: #fff;
-		height: @availableHeight;
-		text-align: center;
-
-		@media @landscape {
-			padding-top: 2px;
-		}
-
-		.title {
-			font-weight: bold;
-			line-height: 1.53;
-			color: #222;
-		}
-
-		.subtitle {
-			font-size: 14px;
-			font-weight: 600;
-			line-height: 1.86;
-			color: #54595d;
-		}
-
-		.message {
-			font-size: 14px;
-			line-height: 1.5;
-			color: #72777d;
-			padding: 10px 40px;
-
-			@media @landscape {
-				padding-left: 16px;
-				padding-right: 16px;
-				margin-bottom: 0;
-				line-height: 1.4;
-				font-size: 13px;
-
-				p {
-					margin-top: 5px;
-					margin-bottom: 4px;
-				}
-			}
-		}
-	}
-}
